+++ changelog.md
@@ -1,3 +1,8 @@
+0.0.1.0
+-------
+
+- use of vector-cereal for cereal instances
+
 0.0.0.4
 
 - stub properties